benteler_frame_supply_agent/pyproject.toml
[project]
name = "benteler-frame-supply-agent"
version = "0.1.0"
description = "A standalone agent for managing Benteler frame supply."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.16",
    "google-adk>=1.8.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.1"
]

[tool.setuptools]
py-modules = ["agent", "agent_executor"]
